- Separate large App.vue file into library and component - Rename components in a clearer manner - Added comments
