+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aab2apk",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
